pyproject.toml
[tool.poetry]
name = "face_rotation_project"
version = "0.1.0"
description = ""
authors = ["Shogo Takata <pineapplehunter.daniel@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
tensorflow = "^2.2.0"
pillow = "^7.1.2"
opencv-python = "^4.2.0"
h5py = "^2.10.0"
matplotlib = "^3.2.1"
numpy = "^1.18.5"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
